feat(exercise): pass fourth test for `signin`
note:
the fourth test is about create new token which should be the same as
`signup` case, but there's an error because of
`{ password: req.body.password }` that should be only `req.body.password`.

also, need to understand when should we use the model, for example
`User`, and the variable `user`.